+++ prototypes/sblgnt-web/utils.py
@@ -1,0 +1,37 @@
+from pysblgnt import morphgnt_rows
+
+
+def rows_by_verses_by_chapters_for_book(book_num):
+
+    last_chapter = 0
+    chapters = []
+    verses = None
+    rows = None
+
+    for row in morphgnt_rows(book_num):
+        chapter = int(row["bcv"][2:4])
+        verse = int(row["bcv"][4:6])
+
+        if chapter != last_chapter:
+
+            if verses:
+                verses[1].append(rows)
+                chapters.append(verses)
+
+            verses = (chapter, [])
+            rows = None
+            last_chapter = chapter
+            last_verse = 0
+
+        if verse != last_verse:
+            if rows:
+                verses[1].append(rows)
+            rows = (verse, [])
+            last_verse = verse
+
+        rows[1].append(row)
+
+    verses[1].append(rows)
+    chapters.append(verses)
+
+    return chapters
